unit test get repositories
